% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knmf_colours.R
\docType{data}
\name{knmf_colours}
\alias{knmf_colours}
\alias{knmf_colors}
\title{Basic colours of ggKNMF package

An eigth-element named vector with rgb codes of basic palette of this package.

First two colours: black and red are colours of KNMF logo. Hues used here have less
luminance to avoid tiring eyes. Indygo and green are triadic-complementary to red used here.
Yellow, orange, blue and pink were generated by algorithmic search for the most distingushible set
(with Polychrome package), were search started from colours adjacent with red and indygo.

Some modifications were made then to obtain better look for color-blind viewers. First six colours
of this palette should be at least moderatly distinguishible for them.

This and other palettes can be previewed in the vignette.}
\format{Named vector with eight elements}
\source{
Harmonic colours were obtained with help of \url{http://paletton.com}
}
\usage{
knmf_colours

knmf_colors
}
\description{
Basic colours of ggKNMF package

An eigth-element named vector with rgb codes of basic palette of this package.

First two colours: black and red are colours of KNMF logo. Hues used here have less
luminance to avoid tiring eyes. Indygo and green are triadic-complementary to red used here.
Yellow, orange, blue and pink were generated by algorithmic search for the most distingushible set
(with Polychrome package), were search started from colours adjacent with red and indygo.

Some modifications were made then to obtain better look for color-blind viewers. First six colours
of this palette should be at least moderatly distinguishible for them.

This and other palettes can be previewed in the vignette.
}
\seealso{
\code{\link{scale_colour_discrete_knmf}}
}
\keyword{datasets}
