% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_colour_continuous_knmf.R
\name{scale_colour_continuous_knmf}
\alias{scale_colour_continuous_knmf}
\alias{monochrome_palette}
\alias{scale_color_continuous_knmf}
\alias{scale_fill_continuous_knmf}
\title{Continuous, monochromatic colour scales

This scale maps values of continuous variables onto monochromatic color scale.
Red and blue colours are supported with default red}
\usage{
monochrome_palette(colour = "red", reverse = FALSE, ...)

scale_colour_continuous_knmf(..., colour = "red", reverse = FALSE)

scale_color_continuous_knmf(..., colour = "red", reverse = FALSE)

scale_fill_continuous_knmf(..., colour = "red", reverse = FALSE)
}
\arguments{
\item{colour}{either "red" or "blue". Choose color for the scale}

\item{reverse}{logical. FALSE (default) if dark colour should represent low values and bright high values
TRUE otherwise}

\item{...}{Arguments passed on to \link[ggplot2]{discrete_scale}}
}
\description{
Continuous, monochromatic colour scales

This scale maps values of continuous variables onto monochromatic color scale.
Red and blue colours are supported with default red
}
\examples{
ggplot(mtcars, aes(hp, mpg, colour = drat)) + geom_point() + scale_colour_continuous_knmf()
ggplot(mtcars, aes(hp, mpg, colour = drat)) + geom_point() +
scale_colour_continuous_knmf(colour = "blue")
}
