% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_colour_discrete_knmf.R
\name{scale_colour_discrete_knmf}
\alias{scale_colour_discrete_knmf}
\alias{knmf_palette}
\alias{scale_color_discrete_knmf}
\alias{scale_fill_discrete_knmf}
\title{Discrete colour scale with ggKNMF basic palette

This scale maps values of categorical variables onto colours from \link{knmf_colours}.
All eight colours should be well distinguishable and first six are colour-blind safe.}
\usage{
knmf_palette(start = 1, reverse = FALSE)

scale_colour_discrete_knmf(..., start = 1, reverse = FALSE,
  na.value = "grey40")

scale_color_discrete_knmf(..., start = 1, reverse = FALSE,
  na.value = "grey40")

scale_fill_discrete_knmf(..., start = 1, reverse = FALSE,
  na.value = "grey40")
}
\arguments{
\item{start}{integer from 1 to 8. From Which value of \link{knmf_colours} should the scale start
taking values. Remember that setting it above one constrains munber of colurs you can use.}

\item{reverse}{logical. If values from \link{knmf_colours} should be taken in reverse order.}

\item{...}{Arguments passed on to \link[ggplot2]{discrete_scale}}

\item{na.value}{Colour for NA}
}
\description{
Discrete colour scale with ggKNMF basic palette

This scale maps values of categorical variables onto colours from \link{knmf_colours}.
All eight colours should be well distinguishable and first six are colour-blind safe.
}
\examples{
ggplot(mtcars, aes(hp, mpg, colour = as.factor(cyl))) + geom_point() +
scale_colour_discrete_knmf()
ggplot(mtcars, aes(x = as.factor(carb), y = ..count.., fill = as.factor(carb))) +
geom_bar() + scale_fill_discrete_knmf()

ggplot(mtcars, aes(hp, fill = as.factor(gear))) +
geom_density(alpha = 0.8) + scale_fill_discrete_knmf()
}
