% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_colour_divergent.R
\name{scale_colour_divergent}
\alias{scale_colour_divergent}
\alias{scale_color_divergent}
\alias{scale_fill_divergent}
\title{Divergent colour scales

To be added}
\usage{
scale_colour_divergent(..., low = "black", mid = "white",
  high = "red", midpoint = 0, na.value = "grey50")

scale_color_divergent(..., low = "black", mid = "white",
  high = "red", midpoint = 0, na.value = "grey50")

scale_fill_divergent(..., low = "black", mid = "white", high = "red",
  midpoint = 0, na.value = "grey50")
}
\arguments{
\item{...}{Arguments passed on to \link[ggplot2]{continuous_scale}}

\item{low}{lower boundary of scale. Name of one of \link{knmf_colours} or corresponding integer}

\item{mid}{colour of midpoint. Any valid colour specyfication.}

\item{high}{upper boundary of scale. Name of one of \link{knmf_colours} or corresponding integer}

\item{midpoint}{to be added}

\item{na.value}{to be added}
}
\description{
Divergent colour scales

To be added
}
