% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_colour_ordinal_knmf.R
\name{scale_colour_ordinal_knmf}
\alias{scale_colour_ordinal_knmf}
\alias{scale_color_ordinal_knmf}
\alias{scale_fill_ordinal_knmf}
\title{Ordinal KNMF colour scale}
\usage{
scale_colour_ordinal_knmf(..., colour = "red", reverse = FALSE)

scale_color_ordinal_knmf(..., colour = "red", reverse = FALSE)

scale_fill_ordinal_knmf(..., colour = "red", reverse = FALSE)
}
\arguments{
\item{...}{arguments passes on to \code{discrete_scale}}

\item{colour}{Dominant colour in scale. Currently \emph{red} and \emph{blue} are supported
with corresponding \code{\link{knmf_colours}}}

\item{reverse}{logical. Sholud oreder of colurs be reversed. As default colours go from dark for low values to
high for high values}
}
\description{
Maps ordered factors to sequence of colours
}
