% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_knmf.R
\name{theme_knmf}
\alias{theme_knmf}
\title{KNMF ggplot theme}
\usage{
theme_knmf(base_size = 11, base_family = "Roboto Condensed",
  base_line_size = base_size/22, base_rect_size = base_size/22)
}
\arguments{
\item{base_size}{base font size. Particular text elements are scaled relatively to this.}

\item{base_family}{base font family. Default is "Roboto Condensed".
If you don't have it installed consider using "sans" value for stable and nice performance.}

\item{base_line_size}{base size for line elements. Particular line elements are scaled relatively to this.}

\item{base_rect_size}{base size for rect elements. Particular rect elements are scaled relatively to this.}
}
\description{
Comlete theme specyfing background, lines, axes, legend and fonts. Default parameters are the same as in
ggplot2 complete themes  like theme_gray, theme_bw, theme_mininal etc.
}
\details{
A simplistic theme with white background, thin axis lines and dashed, grey major grid. It is set as
default when package ggKNMF is loaded.

By default uses Roboto Condensed fonts which must be installed by the user.
}
\examples{
\dontrun{
ggplot(mtcars, aes(hp, fill = as.factor(gear))) + geom_density() +
facet_wrap(~as.factor(am)) +
labs(title = "Density plot", subtitle = "Two facets") + ggKNMF::theme_knmf()
}

}
