% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_knmf_colours.R
\name{use_knmf_colours}
\alias{use_knmf_colours}
\alias{reset_knmf_colours}
\title{KNMF colours}
\usage{
use_knmf_colours(scales = TRUE, geom_defaults = TRUE)

reset_knmf_colours(scales = TRUE, geom_defaults = TRUE)
}
\arguments{
\item{scales}{logical. If TRUE (default) \code{colour} and \code{fill} scales use KNMF colours}

\item{geom_defaults}{logical. If TRUE (default) some standard colour settings are overwritten for KNMF colours
(e.g. outliers in boxplots are red. See Details.)}
}
\description{
Apply colour scheme to scales and deafult some deafult colour values
}
